Refactoring in controller/model methods and previous PR suggestions
All the files listed have been changed to refactor the controller or models methods and merge some files into one. Also, some webpages have been modified according with the previous PR suggestions.
