+++ app/controllers/dashboard_controller.rb
@@ -62,50 +62,27 @@ class DashboardController < MonthlyController
     sign_out_and_redirect(current_user)
   end
 
+  def manage_new_user
+    if @priority.nil?
+      @user.destroy
+    else
+      @priority = 2 if @priority > 2
+      @user.priority = @priority
+      @user.save
+
+      send_out_message(@user.email)
+    end
+    redirect_to controller: 'dashboard', action: 'manager',
+                'report[year]' => fetch_last_year
+  end
+
   private
 
-  def init_overview_options_list
-    @user_options = [['All Salesman', 'All']] +
-                    fetch_username_by_salesman_priority.collect do |username|
-                      [username, username]
-                    end
-    @filter_options = [['Contract Sum', 'CS'], ['Contracts Closed', 'CC'],
-                       ["Goal's Percentage", 'CP']]
-  end
-
-  def verify_filter
-    @filter_options.each do |op|
-      next unless op[1] == params[:report][:goal]
-      @filter_options.delete(op)
-      @filter_options.unshift(op)
-    end
-  end
-
-  def verify_user
-    @user_options.each do |op|
-      next unless op[1] == params[:report][:report_name]
-      @user_options.delete(op)
-      @user_options.unshift(op)
-    end
-  end
-
-  def verify_options
-    verify_filter
-
-    return if params[:report][:report_name] == 'All'
-    verify_user
-  end
-
-  def init_month_year
-    @month_year = params[:report].nil? ? fetch_default_data : fetch_custom_data
-    list = @month_year.split('/')
-
-    @month = list[0]
-    @year = list[1]
-  end
-
-  def fetch_custom_data
-    params[:report][:month] + '/' + params[:report][:year].to_s
+  def send_out_message(email)
+    zapper = ZapierRuby::Zapper.new(:email_zap)
+    zapper.zap(json_maker(email, 'Account permission allowed',
+                          'Your account have been aproved: ' +
+                          ENV['link_to_root']))
   end
 
   def init_manager_data
