+++ app/views/dashboard/manager.html.slim
@@ -8,7 +8,7 @@ body
 
       div.panel-default.panel-heading.form-inline
         h2.h2.font-thin.form-group Spreadsheet Manager
-        = simple_form_for @report, method: :get, url: "/manager", html: { class: "form-group m-l-xl", id: "formSelectBoxYear" } do |f|
+        = simple_form_for @report, method: :get, url: manager_path, html: { class: "form-group m-l-xl", id: "formSelectBoxYear" } do |f|
           = f.input :year, label: false, :collection => @unique_years, :prompt => "Year", input_html: { class: "form-control", id: "monthBox",  :onchange => "sendData('formSelectBoxYear', 1)"}
 
         button.btn.btn-primary.btn-addon.btn-lg.margin.form-group.m-l-md data-target="#exampleModal" data-toggle="modal" type="button"
@@ -22,7 +22,7 @@ body
                 button.close.form-group aria-label="Close" data-dismiss="modal" type="button" onclick="cleanDataSpreadsheet('#{fetch_last_year}')"
                   span aria-hidden="true"
                   | &times;
-              = simple_form_for @report, method: :post, url: "/reports", default: { input_html: { class: "form-control" } } do |f|
+              = simple_form_for @report, method: :post, url: reports_path, default: { input_html: { class: "form-control" } } do |f|
                 .modal-body
                   .form-inline.m-b
                     h4.h4 Spreadsheet name
@@ -62,18 +62,18 @@ body
                 input.numbers hidden="" type="text" value="#{report['year'].to_s}" name="year" id="#{report.id}year"
               td
                 div.form-inline id="#{report.id}divR1"
-                  = simple_form_for @report, method: :get, url: "/monthly_sales", html: { class: "form-group" } do |f|
+                  = simple_form_for @report, method: :get, url: monthly_sales_path, html: { class: "form-group" } do |f|
                     = f.input :month, as: :hidden, input_html: { value: "#{report.month}" }
                     = f.input :year, as: :hidden, input_html: { value: "#{report.year}" }
                     = f.input :report_name, as: :hidden, input_html: { value: "g" }
                     = f.button :submit, "View", class: "btn btn-info m-l"
 
                   button.btn.btn-info.m-l id="#{report.id}editR" onclick="reportVisibility(#{report.id}, true);" Edit
-                  = simple_form_for @report, method: :delete, url: "/reports/"+report.id.to_s, html: { class: "form-group", :onsubmit => "if(!validateMyForm()){event.preventDefault();}" } do |f|
+                  = simple_form_for @report, method: :delete, url: reports_path+"/"+report.id.to_s, html: { class: "form-group", :onsubmit => "if(!validateMyForm()){event.preventDefault();}" } do |f|
                     = f.button :submit, "Delete", class: "btn btn-danger m-l"
 
                 div.form-inline id="#{report.id}divR2" hidden=""
-                  = simple_form_for @report, method: :put, url: "/reports/"+report.id.to_s, html: { class: "form-group", :onsubmit => "if(!validateMyForm()){event.preventDefault();}" } do |f|
+                  = simple_form_for @report, method: :put, url: reports_path+"/"+report.id.to_s, html: { class: "form-group", :onsubmit => "if(!validateMyForm()){event.preventDefault();}" } do |f|
                     = f.input :month_numb, as: :hidden, input_html: { value: "m", id: report.id.to_s+"month"+report.id.to_s }
                     = f.input :year, as: :hidden, input_html: { value: "y", id: report.id.to_s+"year"+report.id.to_s }
                     = f.input :report_name, as: :hidden, input_html: { value: "r", id: report.id.to_s+"report_name"+report.id.to_s }
@@ -105,12 +105,9 @@ body
                   option value="2" Manager
               td
                 div.form-inline
-                  = simple_form_for @user, method: :post, url: "/manage_new_user", html: { class: "form-group", :onsubmit => "if(!validateMyForm()){event.preventDefault();}" } do |f|
+                  = simple_form_for @user, method: :post, url: manage_new_user_path, html: { class: "form-group", :onsubmit => "if(!validateMyForm()){event.preventDefault();}" } do |f|
                     = f.input :priority, as: :hidden, input_html: { value: "p", id: new_user.id.to_s+"permission"+new_user.id.to_s }
                     = f.input :id, as: :hidden, input_html: { value: new_user.id }
                     = f.button :submit, "Allow", class: "btn btn-success m-l", :onclick => "fillUserData(#{new_user.id}, 'yes')"
-
-                  = simple_form_for @user, method: :post, url: "/manage_new_user", html: { class: "form-group", :onsubmit => "if(!validateMyForm()){event.preventDefault();}" } do |f|
-                    = f.input :id, as: :hidden, input_html: { value: new_user.id }
                     = f.button :submit, "Deny", class: "btn btn-danger m-l", :onclick => "fillUserData(#{new_user.id}, 'no')"
