+++ app/views/dashboard/monthly_sales.html.slim
@@ -29,7 +29,7 @@ body onload="generateGraphicPlot(#{@report_points}, #{@contract_points}, #{@day_
               button.close.form-group aria-label="Close" data-dismiss="modal" type="button" onclick="cleanDataContract()"
                 span aria-hidden="true"
                 | &times;
-            = simple_form_for @contract, method: :post, url: "/contracts" do |f|
+            = simple_form_for @contract, method: :post, url: contracts_path do |f|
               .modal-body
                 .form-inline.m-b
                   .form-group
@@ -138,7 +138,7 @@ body onload="generateGraphicPlot(#{@report_points}, #{@contract_points}, #{@day_
                                           - next if user.id == contract.user_id
                                           option value="#{user.id}" = user.full_name
                                     td.v-middle.form-inline
-                                      = simple_form_for @contract, method: :put, url: "/contracts/"+contract.id.to_s, html: { class: "form-group" }, onsubmit: "if(!validateMyForm()){event.preventDefault();}" do |f|
+                                      = simple_form_for @contract, method: :put, url: contracts_path+"/"+contract.id.to_s, html: { class: "form-group" }, onsubmit: "if(!validateMyForm()){event.preventDefault();}" do |f|
                                         = f.input :day, as: :hidden, input_html: { value: "m", id: contract.id.to_s+"day"+contract.id.to_s }
                                         = f.input :store_name, as: :hidden, input_html: { value: "m", id: contract.id.to_s+"store_name"+contract.id.to_s }
                                         = f.input :value, as: :hidden, input_html: { value: "m", id: contract.id.to_s+"value"+contract.id.to_s }
