+++ app/views/user/edit_password.html.slim
@@ -1,6 +1,6 @@
 .wrapper.text-center
   strong Enter your new password
-= simple_form_for @user, method: :post, url: "/users", html: { onsubmit: "if(!verifyPassword()){event.preventDefault();}" } do |f|
+= simple_form_for @user, method: :post, url: user_registration_path, html: { onsubmit: "if(!verifyPassword()){event.preventDefault();}" } do |f|
   .text-danger.text-center id="alertRegister"
     - unless alert.nil?
       - alert.each do |error|
