+++ app/views/user/forgot_password.html.slim
@@ -1,9 +1,9 @@
 .wrapper.text-center
   strong Enter your email to reset your password
-= simple_form_for @user, method: :get, url: "/users/:email" do |f|
+= simple_form_for @user, method: :get, url: user_registration_path+"/:email" do |f|
   .list-group.list-group-sm
     .list-group-item
       = f.input :email, as: :email, label: false, input_html: { class: "form-control no-border", placeholder:"Email" }, wrapper: false
   = f.button :submit, "Send", class: "btn btn-lg btn-primary btn-block"
 .text-center.m-t.m-b
-  = link_to 'I remeber my password', root_path
+  = link_to 'I remember my password', root_path
