+++ app/views/user/register.html.slim
@@ -1,4 +1,4 @@
-= simple_form_for @user, method: :post, url: "/users", html: { onsubmit: "if(!verifyPassword()){event.preventDefault();}" } do |f|
+= simple_form_for @user, method: :post, url: user_registration_path, html: { onsubmit: "if(!verifyPassword()){event.preventDefault();}" } do |f|
   .text-center.m-t-md
     h3.text-center Sign Up
     strong Please, fill in all the data below
